+++ src/search/data_loader.h
@@ -58,8 +58,8 @@ struct SearchEntity
 	uint8  nation;
 	uint8  rank;
 	uint8  race;
-	uint8  zone;
-	uint8  prevzone;
+	uint16 zone;
+	uint16 prevzone;
 	uint16 flags1;
 	uint32 flags2;
 	uint32 comment;
